pyproject.toml
[project]
name = "argentic"
version = "0.11.2"
description = ""
authors = [{ name = "Iurii", email = "angkira0@gmail.com" }]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
keywords = [
    "ai",
    "ai agent",
    "rag",
    "ollama",
    "sentence-transformers",
    "chromadb",
]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "ollama>=0.4.8,<0.5",
    "chromadb>=1.0.9,<2",
    "sentence-transformers>=4.1.0,<5",
    "pyyaml>=6.0.2,<7",
    "requests>=2.31.0,<3",
    "aiomqtt>=2.4.0",
    "google-generativeai>=0.8.5",
    "python-dotenv",
    "tenacity>=9.0.0,<10",
    "yaspin>=3.2.0",
]

[project.scripts]
argentic = "argentic.__main__:main"

[project.optional-dependencies]
dev = [
    "black>=25.1.0,<26",
    "isort>=6.0.1,<7",
    "pytest>=8.4.1,<9",
    "pytest-asyncio>=0.23.5,<0.24",
    "python-on-whales>=0.70.0",
    "flake8>=7.2.0",
    "commitizen>=4.8.3,<5",
    "pre-commit>=4.2.0,<5",
    "pyrefly>=0.15.2",
    "build>=1.2.2.post1",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.6.14",
    "pymdown-extensions>=10.0",
    "mkdocstrings[python]>=0.24",
    "mkdocs-same-dir>=0.1.3",
    "mkdocs-macros-plugin>=1.3.7"
]
kafka = [
    "aiokafka>=0.12.0"
]
redis = [
    "aioredis>=2.0.1"
]
rabbitmq = [
    "aio-pika>=9.5.5"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/argentic"]

[tool.black]
line-length = 100

[tool.hatch.env.default]
variables = { PYTHONPATH = "src" }

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
max_line_length = 150

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver2"
version_provider = "pep621"
version_files = ["pyproject.toml:version"]
update_changelog_on_bump = false
major_version_zero = true

[tool.pyrefly]
project_includes = ["**/*"]
project_excludes = [
    "**/.[!/.]*",
    "**/*venv/**/*",
    "**/*__pycache__/**/*",
    "**/tests/**/*",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]
